Migrate away from displaying bias and weights
The rational for this is that these values can (and often will)
be quite long and won't fit in the bubble. We also do not
want to truncate these values. Instead, these are displayed in the
input panel, and IDS and input indicies are displayed on the nodes
and edges.